+++ renderer/components/index.ts
@@ -42,7 +42,6 @@ export { SettingsBackup } from './SettingsBackup';
 export { AppStatistics } from './AppStatistics';
 export { SettingsSidebar } from './SettingsSidebar';
 export type { SettingsCategory } from './SettingsSidebar';
-export { SeedingDashboard } from './SeedingDashboard';
 export { PrivacySettings } from './PrivacySettings';
 export { TorrentControlModal } from './TorrentControlModal';
 export { StreamPlayerModal } from './StreamPlayerModal';